+++ client/src/index.js
@@ -5,7 +5,6 @@ import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import { Auth0Provider } from "./react-auth0-spa";
 import history from "./utils/history";
-import config from "./auth_config.json";
 // A function that routes the user to the right place
 // after login
 const onRedirectCallback = appState => {
